requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.30
